reworking testrunner to be more stable

+++ TestSystem/libs/ServerResource.py
@@ -14,36 +14,62 @@ import subprocess
 class ServerResource (TestResource):
 
 	def reader(self):
-		while True:
-			try:
+	
+		try:
+			while True:
 				line = self.p.stdout.readline().decode("utf-8")
 				#print("\"" + str(line) + "\"")
 				self.queue.put(line)
 
-			except (KeyboardInterrupt, SystemExit):
-				self.deallocate()
+		except (KeyboardInterrupt, SystemExit):
+			print("server readThread exit")
+			self.deallocate()
+		except Exception as e:
+			print(e)
+			self.readFail = True
+			
 			
 
 	def __init__(self, host, port):
+		print("init server")
 		self.host = host
 		self.port = port
-		self.p = None
-		self.connect()
-		self.queue = Queue()	
-		self.recThread = Thread(target = self.reader, args=())
-		self.recThread.setDaemon( True)
-		self.recThread.start();
-
 		
-
-
-	def deallocate(self):
-		print("dealocate server")
+		self.p = None
+		self.queue = None
+		self.readFail = False
+		self.reset()
+	
+	def close(self):
 		try:
-			self.p.kill()
+			if(self.p != None):
+				self.p.kill()
+			self.p = None	
 		except Exception as e:
 			print(e)
 			pass
+		
+			
+	def reset(self):
+		print("reset server")
+		time.sleep(1)
+		try:
+			if(self.p != None):
+				self.p.kill()
+
+			self.queue = Queue()
+			self.p = None	
+			self.connect()
+			
+			self.recThread = Thread(target = self.reader, args=())
+			self.recThread.setDaemon( True)
+			self.recThread.start();
+		except KeyboardInterrupt:
+			pass
+		except Exception as e:
+			print(e)
+			pass
+
 
 	def connect(self):
 		try:
@@ -52,19 +78,28 @@ class ServerResource (TestResource):
 			pass
 		except Exception as e:
 			print(e)
+			self.reset()
 
 	def send(self, message):
-		totaltsent = 0
-		##print("sent="+ str(message))
-		##self.p.communicate(message)
-		#print("done")
-		print("Server << " + message)
+		try:
+			if(self.readFail == False):
+				totaltsent = 0
+				##print("sent="+ str(message))
+				##self.p.communicate(message)
+				#print("done")
+				print("Server << " + message)
+				return
+		except (KeyboardInterrupt):
+			return
+		except Exception as e:
+			print(e)
+		raise Exception("Test Failed Server message " + str(message) + " not sent on server")
 
 	def recv(self, regX, timeout):
 		try:	
 			r = re.compile(regX)
 			end = time.time() + timeout
-			while( end > time.time()):
+			while( end > time.time() and self.readFail == False):
 				try: 
 					line = ( self.queue.get_nowait())
 					print("Server >> " + str(line.replace("\n","")))
@@ -77,6 +112,8 @@ class ServerResource (TestResource):
 					continue
 		except (KeyboardInterrupt):
 			return False
+		except Exception as e:
+			print(e)
 
 		raise Exception("Test Failed Server message " + str(regX) + " not received on server")
 		return False
